Move Rocket, MiniRocket, and MiniRocketMultivariate to submodule
