+++ sktime/transformations/panel/rocket/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ["Rocket", "MiniRocket", "MiniRocketMultivariate"]
+
+from ._rocket import Rocket
+from ._minirocket import MiniRocket
+from ._minirocket_multivariate import MiniRocketMultivariate
